refactor(#102): extract video-analytics data layer shared by CLI and MCP
Phase 1 of the structured-returns migration: lib/analytics.ts owns the
video-analytics business logic (video lookup, date-range defaulting,
dimension validation, 90-day chunking, row aggregation) exactly once.
The get-video-analytics command keeps its spinner and five output
formats on top; the MCP youtube_get_video_analytics tool consumes the
typed result directly — dropping its 70-line inline copy and gaining
the dimensions support (#120) and video title that copy never had.
